Restructuring for the Mediaflux::Http Module
