+++ app/models/mediaflux/http/collection_list_request.rb
@@ -1,0 +1,20 @@
+# frozen_string_literal: true
+module Mediaflux
+  module Http
+    class CollectionListRequest < Request
+      def self.default_namespace
+        "tigerdata"
+      end
+
+      def self.service
+        "asset.collection.list"
+      end
+
+      def self.service_args
+        {
+          namespace: default_namespace
+        }
+      end
+    end
+  end
+end
